question controller

+++ app/Http/Controllers/QuestionController.php
@@ -2,11 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\QuestionRequest;
 use App\Models\Course;
 use App\Models\Question;
 use App\Models\QuestionOption;
 use App\Models\Test;
-use Illuminate\Http\Request;
+use Illuminate\Database\Eloquent\Collection;
 
 class QuestionController extends Controller
 {
@@ -19,14 +20,9 @@ class QuestionController extends Controller
         ]);
     }
 
-    public function store(Request $request, Course $course, Test $test)
+    public function store(QuestionRequest $request, Course $course, Test $test)
     {
-        // dd($request->all());
-        // $request->validate([
-        //     'name' => 'required',
-        //     'answer' => 'required',
-        //     'is_answer' => 'required'
-        // ]);
+        $request->validated();
 
         if ($request->input('is_answer') == 'answer1') {
             $is_answer = [1, 0];
@@ -40,35 +36,44 @@ class QuestionController extends Controller
 
         $test->questions()->attach($question);
 
-        foreach ($request->input('answer') as $index => $answer) {
-            QuestionOption::create([
+        $collection = new Collection($request->answer);
+
+        $collection->each(function ($answer, $key) use ($question, $is_answer) {
+            $questionOption = QuestionOption::make([
                 'question_id' => $question->id,
                 'name' => $answer,
-                'is_answer' => $is_answer[$index]
+                'is_answer' => $is_answer[$key]
             ]);
-        }
+            $questionOption->save();
+        });
+
         if ($request->input('action') === 'save') {
-            return redirect()->route('courses.tests.questions.edit', [$course, $test, $question])->with('success', 'question created successfully');
+            return redirect()->route('courses.tests.questions.edit', [$course, $test, $question])
+                ->with('success', __('question created successfully'));
         }
 
-        return back()->with('success', 'question created successfully');
+        return back()->with('success', __('question created successfully'));
     }
 
     public function edit(Course $course, Test $test, Question $question)
     {
-        $questionOption = $question->questionOptions()->get();
+        $this->authorize('edit', $course);
 
         return view('questions.edit', [
             'course' => $course,
             'unit' => $test->lesson->load('unit')->unit,
             'test' => $test,
             'question' => $question,
-            'questionOption' => $questionOption
+            'questionOption' => $question->questionOptions()->get()
         ]);
     }
 
-    public function update(Request $request, Course $course, Test $test, Question $question)
+    public function update(QuestionRequest $request, Course $course, Test $test, Question $question)
     {
+        $this->authorize('update', $course);
+
+        $request->validated();
+
         if ($request->input('is_answer') == 'answer1') {
             $is_answer = [1, 0];
         } elseif ($request->input('is_answer') == 'answer2') {
@@ -79,24 +84,26 @@ class QuestionController extends Controller
             'name' => $request->name,
         ]);
 
-        foreach ($request->input('answer') as $index => $answer) {
-            $question->questionOptions[$index]->update([
-                'question_id' => $question->id,
-                'name' => $answer,
-                'is_answer' => $is_answer[$index]
-            ]);
-        }
+        $question->questionOptions->each(function ($answer, $key) use ($request, $question, $is_answer) {
 
-        return back()->with('success', 'question updated successfully');
+            $question->questionOptions[$key]->name = $request->answer[$key];
+            $question->questionOptions[$key]->is_answer = $is_answer[$key];
+
+            $question->questionOptions[$key]->save();
+        });
+
+        return back()->with('success',  __('question updated successfully'));
     }
 
     public function destroy(Course $course, Test $test, Question $question)
     {
+        $this->authorize('destroy', $course);
+
         $test->questions()->detach($question);
 
         $question->questionOptions()->delete();
         $question->delete();
 
-        return back()->with('success', 'question deleted successfully');
+        return back()->with('success',  __('question deleted successfully'));
     }
 }
